Notes entre 0 et 5 avec pas de 0.5

+++ symfony/src/Entity/Rating.php
@@ -80,7 +80,7 @@ class Rating
         return $this->value;
     }
 
-    public function setValue(int $value): self
+    public function setValue(float $value): self
     {
         $this->value = $value;
